feat(gap): implement gap filler task
- GapFiller is a task that queries the visor_gap_reports table for gaps
with status `GAP` to process. GapFiller runs a walk task for
each gap it finds in the visor_gap_reports table. GapFiller accepts a
slice of task that may be used to filter that types of gaps filled. When
the GapFiller successfully fills a gap it updates its `status` in the
visor_gap_reports table from `GAP` to `FILLED`. If the GapFiller fails
to walk the gap the error causing the failuer is written to the `status`
column.